+++ test/unit/fuzz-tools.test.js
@@ -432,6 +432,7 @@ import * as brookerV452 from '../../lib/brooker-v452.js';
 import * as badoV454 from '../../lib/bado-v454.js';
 import * as nunleyVertulloV455 from '../../lib/nunley-vertullo-v455.js';
 import * as leddyPackerV456 from '../../lib/leddy-packer-v456.js';
+import * as stulbergV457 from '../../lib/stulberg-v457.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -809,6 +810,7 @@ const MODULES = {
   'bado-v454.js': badoV454,
   'nunley-vertullo-v455.js': nunleyVertulloV455,
   'leddy-packer-v456.js': leddyPackerV456,
+  'stulberg-v457.js': stulbergV457,
   'tb-testing.js': tbTesting,
 };